+++ src/command/delete_database.rs
@@ -54,15 +54,9 @@ impl<'a, P: IntoDatabasePath> Command for DeleteDatabase<'a, P> {
             hyper::status::StatusCode::Ok => {
                 command::content_type_must_be_application_json(&resp.headers)
             }
-            hyper::status::StatusCode::BadRequest => {
-                Err(Error::BadRequest(try!(json::decode_json::<_, ErrorResponse>(resp))))
-            }
-            hyper::status::StatusCode::Unauthorized => {
-                Err(Error::Unauthorized(Some(try!(json::decode_json::<_, ErrorResponse>(resp)))))
-            }
-            hyper::status::StatusCode::NotFound => {
-                Err(Error::NotFound(Some(try!(json::decode_json::<_, ErrorResponse>(resp)))))
-            }
+            hyper::status::StatusCode::BadRequest => Err(make_couchdb_error!(BadRequest, resp)),
+            hyper::status::StatusCode::Unauthorized => Err(make_couchdb_error!(Unauthorized, resp)),
+            hyper::status::StatusCode::NotFound => Err(make_couchdb_error!(NotFound, resp)),
             _ => Err(Error::UnexpectedHttpStatus { got: resp.status }),
         }
     }