+++ src/command/get_view.rs
@@ -133,17 +133,11 @@ impl<'a, P, K, V> Command for GetView<'a, P, K, V>
                 let view_result = try!(json::decode_json::<_, ViewResult<K, V>>(resp));
                 Ok(view_result)
             }
-            hyper::status::StatusCode::BadRequest => {
-                Err(Error::BadRequest(try!(json::decode_json::<_, ErrorResponse>(resp))))
-            }
-            hyper::status::StatusCode::Unauthorized => {
-                Err(Error::Unauthorized(Some(try!(json::decode_json::<_, ErrorResponse>(resp)))))
-            }
-            hyper::status::StatusCode::NotFound => {
-                Err(Error::NotFound(Some(try!(json::decode_json::<_, ErrorResponse>(resp)))))
-            }
+            hyper::status::StatusCode::BadRequest => Err(make_couchdb_error!(BadRequest, resp)),
+            hyper::status::StatusCode::Unauthorized => Err(make_couchdb_error!(Unauthorized, resp)),
+            hyper::status::StatusCode::NotFound => Err(make_couchdb_error!(NotFound, resp)),
             hyper::status::StatusCode::InternalServerError => {
-                Err(Error::InternalServerError(try!(json::decode_json::<_, ErrorResponse>(resp))))
+                Err(make_couchdb_error!(InternalServerError, resp))
             }
             _ => Err(Error::UnexpectedHttpStatus { got: resp.status }),
         }