+++ src/command/put_document.rs
@@ -82,18 +82,10 @@ impl<'a, P: IntoDocumentPath, T: 'a + serde::Serialize> Command for PutDocument<
                 let rev: Revision = content.rev.into();
                 Ok(rev)
             }
-            hyper::status::StatusCode::BadRequest => {
-                Err(Error::BadRequest(try!(json::decode_json::<_, ErrorResponse>(resp))))
-            }
-            hyper::status::StatusCode::Unauthorized => {
-                Err(Error::Unauthorized(Some(try!(json::decode_json::<_, ErrorResponse>(resp)))))
-            }
-            hyper::status::StatusCode::NotFound => {
-                Err(Error::NotFound(Some(try!(json::decode_json::<_, ErrorResponse>(resp)))))
-            }
-            hyper::status::StatusCode::Conflict => {
-                Err(Error::DocumentConflict(try!(json::decode_json::<_, ErrorResponse>(resp))))
-            }
+            hyper::status::StatusCode::BadRequest => Err(make_couchdb_error!(BadRequest, resp)),
+            hyper::status::StatusCode::Unauthorized => Err(make_couchdb_error!(Unauthorized, resp)),
+            hyper::status::StatusCode::NotFound => Err(make_couchdb_error!(NotFound, resp)),
+            hyper::status::StatusCode::Conflict => Err(make_couchdb_error!(DocumentConflict, resp)),
             _ => Err(Error::UnexpectedHttpStatus { got: resp.status }),
         }
     }